move NgTriplet.java file position to main project
